+++ tools/build.js
@@ -2,18 +2,24 @@ var fs = require('fs');
 var path = require('path');
 var stringify = require('json-stable-stringify');
 var commandPath = path.join(__dirname, '..', 'commands.json');
+var redisCommands = require('../');
 
 var Redis = require('ioredis');
 var redis = new Redis(process.env.REDIS_URI);
 
-redis.command(function (err, res) {
+redis.command().then(function (res) {
   redis.disconnect();
 
-  if (err) {
-    throw err;
-  }
+  // Find all special handled cases
+  var movableKeys = String(redisCommands.getKeyIndexes).match(/case '[a-z-]+':/g).map(function (entry) {
+    return entry.replace(/^case \'|\':$/g, '');
+  });
 
   var commands = res.reduce(function (prev, current) {
+    var currentCommandPos = movableKeys.indexOf(current[0]);
+    if (currentCommandPos !== -1 && current[2].indexOf('movablekeys') !== -1) {
+      movableKeys.splice(currentCommandPos, 1);
+    }
     // https://github.com/antirez/redis/issues/2598
     if (current[0] === 'brpop' && current[4] === 1) {
       current[4] = -2;
@@ -42,6 +48,10 @@ redis.command(function (err, res) {
       keyStop: 0,
       step: 0
     }
+  }
+
+  if (movableKeys.length !== 0) {
+    throw new Error('Not all commands (\'' + movableKeys.join('\', \'') + '\') with the "movablekeys" flag are handled in the code');
   }
 
   // Use json-stable-stringify instead fo JSON.stringify